TEXTURES WORK!

+++ src/engine/entity/component/SpriteComponent.java
@@ -10,12 +10,12 @@ public class SpriteComponent extends AbstractComponent
 {
 	private Renderable2D m_Sprite;
 	
-	public SpriteComponent(Vec4 color, String textureImage, int width, int height)
+	public SpriteComponent(Vec4 color, String textureImage, float width, float height)
 	{
 		m_Sprite = new Renderable2D(width, height, color, textureImage);
 	}
 	
-	public SpriteComponent(Vec4 color, int width, int height)
+	public SpriteComponent(Vec4 color, float width, float height)
 	{
 		m_Sprite = new Renderable2D(width, height, color);
 	}
